browserauto/requirements.txt
# Core Framework
Django==5.2.8
python-dotenv==1.0.0

# LangChain & LangGraph - AI Agent Framework
langchain==0.3.21
langchain-core==1.0.4
langchain-google-genai==2.0.8
langgraph==1.0.4
langgraph-swarm==0.2.0
langgraph-checkpoint==2.0.9

# Google Gemini AI
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1

# Computer Vision & Automation
PyAutoGUI==0.9.54
Pillow==12.0.0
opencv-python==4.12.0.88
numpy==2.2.6

# Web Automation (Optional/Legacy)
selenium==4.27.1

# Supporting Libraries
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
sniffio==1.3.1

# JSON Processing
jsonpatch==1.33
jsonpointer==3.0.0

# Type Checking & Validation
typing-extensions==4.12.2
annotated-types==0.7.0
pydantic==2.10.6
pydantic-core==2.27.2

# Async & Concurrency
asyncio
aiohttp==3.11.11

# Utilities
filetype==1.2.0
distro==1.9.0
colorama==0.4.6
cachetools==6.2.1

# gRPC & Protocol Buffers
grpcio==1.76.0
grpcio-status==1.71.2
googleapis-common-protos==1.72.0
protobuf==5.29.3

# HTTP Libraries
httplib2==0.31.0
requests==2.32.3
urllib3==2.3.0

# YAML Support
PyYAML==6.0.2

# Database (Django default)
asgiref==3.10.0
sqlparse==0.5.3

# Jinja2 for templating
jinja2==3.1.5
MarkupSafe==3.0.2

# Other utilities
tenacity==9.0.0
orjson==3.10.14
packaging==24.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
rsa==4.9
